[core] quiet compiler warnings
cast away signedness warning in request_check_hostname()
mod_gnutls https_add_ssl_client_entries crts allocated if crt_size != 0
  (which is already checked earlier in routine)

report from FaceBook Infer static analysis tool (https://fbinfer.com/)
- quiet dead store warnings
- check return != NULL from allocation funcs

+++ src/request.c
@@ -36,7 +36,7 @@ static int request_check_hostname(buffer * const host) {
     if (*h != '[') {
         uint32_t len = buffer_clen(host);
         const char * const colon = memchr(h, ':', len);
-        uint32_t hlen = colon ? (colon - h) : len;
+        uint32_t hlen = colon ? (uint32_t)(colon - h) : len;
 
         /* if hostname ends in ".", strip it */
         if (__builtin_expect( (0 == hlen), 0)) return -1;
